refactor: 함수들을 modules 폴더의 index.js 파일로 분리
