+++ src/App.js
@@ -1,70 +1,15 @@
-const { Console, Random } = require('@woowacourse/mission-utils');
+const { Console } = require('@woowacourse/mission-utils');
+
+const { GAME_VALUE, GAME_MESSAGE } = require('./assets/constants');
 const {
-  GAME_VALUE,
-  GAME_MESSAGE,
-  ERROR_MESSAGE,
-} = require('./assets/constants');
-
-const isValidRestartInputValue = (inputValue) => {
-  if (!['1', '2'].includes(inputValue)) {
-    throw ERROR_MESSAGE.invalidRestartValue;
-  }
-};
-
-const getGameResultMessage = ({ strike, ball }) => {
-  if (strike === 0 && ball === 0) {
-    return '낫싱';
-  }
-  if (strike && ball) {
-    return `${ball}볼 ${strike}스트라이크`;
-  }
-  if (strike) {
-    return `${strike}스트라이크`;
-  }
-  if (ball) {
-    return `${ball}볼`;
-  }
-};
-
-const getSameNumCount = (userNumbers, computerNumbers) => {
-  return [...userNumbers].filter((num) => computerNumbers.includes(num)).length;
-};
-
-const getStrikeCount = (userNumbers, computerNumbers) => {
-  return [...userNumbers].filter((num, i) => num === computerNumbers[i]).length;
-};
-
-const haveSameNumber = (userNumbers) => {
-  if (new Set([...userNumbers]).size < 3) {
-    throw ERROR_MESSAGE.duplicateError;
-  }
-};
-
-const isValidRangeNumber = (userNumbers) => {
-  const regex = /^[1-9]{3}$/;
-  if (!regex.test(userNumbers)) {
-    throw ERROR_MESSAGE.invalidValueError;
-  }
-};
-
-const getNotContainNumber = (randomNumbers, number) => {
-  if (randomNumbers.includes(number)) {
-    return [];
-  }
-
-  return [number];
-};
-
-const getRandomNumbers = (size, start, end) => {
-  let randomNumbers = [];
-  while (randomNumbers.length < size) {
-    const number = Random.pickNumberInRange(start, end);
-    const notContainNumber = getNotContainNumber(randomNumbers, number);
-    randomNumbers = [...randomNumbers, ...notContainNumber];
-  }
-
-  return randomNumbers.join('');
-};
+  isValidRestartInputValue,
+  getGameResultMessage,
+  getSameNumCount,
+  getStrikeCount,
+  haveSameNumber,
+  isValidRangeNumber,
+  getRandomNumbers,
+} = require('./modules/index');
 
 class App {
   constructor() {
